Fixed the issue where the vehicle could not reverse when the speed was too slow. Fixed the oscillation issue when reversing.
