+++ GEMstack/onboard/planning/stanley.py
@@ -1,8 +1,6 @@
-###supports purely forward routes###
 import numpy as np
 from math import sin, cos, atan2
 
-# These imports match your existing project structure
 from ...mathutils.control import PID
 from ...utils import settings
 from ...mathutils import transforms
@@ -28,15 +26,6 @@ def normalise_angle(angle):
 # 2. Stanley-based controller with longitudinal PID
 #####################################
 class Stanley(object):
-    """
-    A Stanley controller that handles lateral control (steering)
-    plus a basic longitudinal control (PID + optional feedforward).
-    It has been modified to reduce oscillations by:
-      1) Lower gains
-      2) Steering damping
-      3) Low-pass filter on steering
-      4) Gentler speed logic when cornering
-    """
 
     def __init__(
         self,
@@ -47,31 +36,24 @@ class Stanley(object):
         """
         :param control_gain:       Stanley lateral control gain k (lowered from the default to reduce overshoot).
         :param softening_gain:     Softening gain k_soft.
-        :param yaw_rate_gain:      Yaw-rate gain k_yaw_rate (helps reduce oscillations).
-        :param steering_damp_gain: Steering damping gain k_damp_steer.
         :param desired_speed:      Desired speed (float) or 'path'/'trajectory'.
         """
-
-        # 1) Lower Gains
         self.k = control_gain if control_gain is not None else settings.get('control.stanley.control_gain')
         self.k_soft = softening_gain if softening_gain is not None else settings.get('control.stanley.softening_gain')
 
-        # Typically, this is the max front-wheel steering angle in radians
         self.max_steer = settings.get('vehicle.geometry.max_wheel_angle')
         self.wheelbase = settings.get('vehicle.geometry.wheelbase')
 
-        # Basic longitudinal constraints
         self.speed_limit = settings.get('vehicle.limits.max_speed')
-        self.max_accel   = settings.get('vehicle.limits.max_acceleration')
-        self.max_decel   = settings.get('vehicle.limits.max_deceleration')
+        self.reverse_speed_limit = settings.get('vehicle.limits.max_reverse_speed')
+        self.max_accel = settings.get('vehicle.limits.max_acceleration')
+        self.max_decel = settings.get('vehicle.limits.max_deceleration')
 
-        # PID for longitudinal speed tracking
         p = settings.get('control.longitudinal_control.pid_p')
         d = settings.get('control.longitudinal_control.pid_d')
         i = settings.get('control.longitudinal_control.pid_i')
         self.pid_speed = PID(p, d, i, windup_limit=20)
-
-        # Speed source: numeric or derived from path/trajectory
+        
         if desired_speed is not None:
             self.desired_speed_source = desired_speed
         else:
@@ -82,25 +64,23 @@ class Stanley(object):
         else:
             self.desired_speed = None
 
-        # For path/trajectory
         self.path_arg = None
         self.path = None
         self.trajectory = None
         self.current_path_parameter = 0.0
         self.current_traj_parameter = 0.0
         self.t_last = None
+        self.reverse = None
+        self.sharp_turn = False
 
     def set_path(self, path: Path):
-        """Sets the path or trajectory to track."""
         if path == self.path_arg:
             return
         self.path_arg = path
 
-        # If the path has more than 2D, reduce to (x,y)
         if len(path.points[0]) > 2:
             path = path.get_dims([0,1])
 
-        # If no timing info, we can't rely on 'path'/'trajectory' speed
         if not isinstance(path, Trajectory):
             self.path = path.arc_length_parameterize()
             self.trajectory = None
@@ -120,14 +100,116 @@ class Stanley(object):
         fy = y + self.wheelbase * sin(yaw)
         return fx, fy
 
+    def _find_rear_axle_position(self, x, y, yaw):
+        """Compute rear-axle world position from the vehicle's reference point and yaw."""
+        rx = x - self.wheelbase * cos(yaw)
+        ry = y - self.wheelbase * sin(yaw)
+        return rx, ry
+
+    def initialize_state_and_direction(self, state: VehicleState):
+        if self.path is None:
+            raise ValueError("Stanley: Path must be set before initializing state and direction.")
+        curr_x = state.pose.x
+        curr_y = state.pose.y
+        curr_yaw = state.pose.yaw if state.pose.yaw is not None else 0.0
+        fx, fy = self._find_front_axle_position(curr_x, curr_y, curr_yaw)
+        path_domain_start, path_domain_end = self.path.domain()
+        search_end = min(path_domain_end, path_domain_start + 5.0)
+        search_domain_start = [path_domain_start, search_end]
+        _, closest_param_at_start = self.path.closest_point_local((fx, fy), search_domain_start)
+        tangent_at_start = self.path.eval_tangent(closest_param_at_start)
+        initial_reverse = False
+        if np.linalg.norm(tangent_at_start) > 1e-6:
+            path_yaw_at_start = atan2(tangent_at_start[1], tangent_at_start[0])
+            heading_diff = normalise_angle(path_yaw_at_start - curr_yaw)
+            initial_reverse = abs(heading_diff) > (np.pi / 2.0)
+        self.pid_speed.reset()
+        self.current_path_parameter = closest_param_at_start
+        if self.trajectory:
+            self.current_traj_parameter = self.trajectory.domain()[0]
+        return initial_reverse
+
+    def is_target_behind_vehicle(self, vehicle_pose, target_point_coords):
+        curr_x = vehicle_pose.x
+        curr_y = vehicle_pose.y
+        curr_yaw = vehicle_pose.yaw
+        if curr_yaw is None:
+            return False
+        target_x = target_point_coords[0]
+        target_y = target_point_coords[1]
+        vec_x = target_x - curr_x
+        vec_y = target_y - curr_y
+        if abs(vec_x) < 1e-6 and abs(vec_y) < 1e-6:
+            return False
+        heading_x = cos(curr_yaw)
+        heading_y = sin(curr_yaw)
+        dot_product = vec_x * heading_x + vec_y * heading_y
+        return (dot_product < 0)
+
+    def _check_sharp_turn_ahead(self, lookahead_s=3.0, threshold_angle=np.pi/2.0, num_steps=4):
+        if not self.path:
+            return False
+
+        path = self.path
+        current_s = self.current_path_parameter
+        domain_start, domain_end = path.domain()
+
+        if current_s >= domain_end - 1e-3:
+            return False
+
+        step_s = lookahead_s / num_steps
+        s_prev = current_s
+        
+        try:
+            tangent_prev = path.eval_tangent(s_prev)
+            if np.linalg.norm(tangent_prev) < 1e-6:
+                s_prev_adjusted = min(s_prev + step_s / 2, domain_end)
+                if s_prev_adjusted <= s_prev:
+                    return False
+                tangent_prev = path.eval_tangent(s_prev_adjusted)
+                if np.linalg.norm(tangent_prev) < 1e-6:
+                    return False
+                s_prev = s_prev_adjusted
+                
+            angle_prev = atan2(tangent_prev[1], tangent_prev[0])
+
+        except Exception as e:
+            return False
+
+        for i in range(num_steps):
+            s_next = s_prev + step_s
+            s_next = min(s_next, domain_end)
+
+            if s_next <= s_prev + 1e-6:
+                break
+
+            try:
+                tangent_next = path.eval_tangent(s_next)
+                if np.linalg.norm(tangent_next) < 1e-6:
+                    s_prev = s_next
+                    continue 
+
+                angle_next = atan2(tangent_next[1], tangent_next[0])
+            except Exception as e:
+                break
+
+            angle_change = abs(normalise_angle(angle_next - angle_prev))
+
+            if angle_change > threshold_angle:
+                return True
+
+            angle_prev = angle_next
+            s_prev = s_next
+        return False
+
     def compute(self, state: VehicleState, component: Component = None):
-        """Compute the control outputs: (longitudinal acceleration, front wheel angle)."""
         t = state.pose.t
         if self.t_last is None:
             self.t_last = t
-        dt = t - self.t_last
+            dt = 0
+        else:
+            dt = t - self.t_last
 
-        # Current vehicle states
         curr_x = state.pose.x
         curr_y = state.pose.y
         curr_yaw = state.pose.yaw if state.pose.yaw is not None else 0.0
@@ -136,9 +218,10 @@ class Stanley(object):
         if self.path is None:
             if component:
                 component.debug_event("No path provided to Stanley controller. Doing nothing.")
+            self.t_last = t
+            self.pid_speed.reset()
             return (0.0, 0.0)
 
-        # Ensure same frame
         if self.path.frame != state.pose.frame:
             if component:
                 component.debug_event(f"Transforming path from {self.path.frame.name} to {state.pose.frame.name}")
@@ -149,91 +232,160 @@ class Stanley(object):
                 component.debug_event(f"Transforming trajectory from {self.trajectory.frame.name} to {state.pose.frame.name}")
             self.trajectory = self.trajectory.to_frame(state.pose.frame, current_pose=state.pose)
 
-        # 1) Closest point
-        fx, fy = self._find_front_axle_position(curr_x, curr_y, curr_yaw)
-        search_start = self.current_path_parameter - 5.0
+
+        if self.reverse is None:
+            self.reverse = self.initialize_state_and_direction(state)
+
+        if self.reverse:
+            fx, fy = self._find_rear_axle_position(curr_x, curr_y, curr_yaw)
+        else:
+            fx, fy = self._find_front_axle_position(curr_x, curr_y, curr_yaw)
+        search_start = self.current_path_parameter
         search_end   = self.current_path_parameter + 5.0
         closest_dist, closest_parameter = self.path.closest_point_local((fx, fy), [search_start, search_end])
         self.current_path_parameter = closest_parameter
 
-        # 2) Path heading
-        target_x, target_y = self.path.eval(closest_parameter)
-        tangent = self.path.eval_tangent(closest_parameter)
+        target_x, target_y = self.path.eval(self.current_path_parameter)
+        tangent = self.path.eval_tangent(self.current_path_parameter)
         path_yaw = atan2(tangent[1], tangent[0])
-        desired_x = target_x
-        desired_y = target_y
-        # 3) Lateral error
+
         dx = fx - target_x
         dy = fy - target_y
-        left_vec = np.array([sin(curr_yaw), -cos(curr_yaw)])
-        cross_track_error = np.sign(np.dot(np.array([dx, dy]), left_vec)) * closest_dist
 
-        # 4) Heading error
+        if not self.reverse:
+            try:
+                is_sharp_turn_ahead = self._check_sharp_turn_ahead(
+                    lookahead_s=2.0,     
+                    threshold_angle=np.pi/2.0
+                )
+            except Exception as e:
+                is_sharp_turn_ahead = False
+            if is_sharp_turn_ahead and not self.sharp_turn:
+                self.sharp_turn = True
+            use_reverse = self.is_target_behind_vehicle(state.pose, (target_x, target_y))
+            if use_reverse and self.sharp_turn:
+                self.reverse = True
+                self.sharp_turn = False
+        
+        if self.reverse:
+            cross_track_error = dx * (-tangent[1]) + dy * tangent[0]
+        else:
+            left_vec = np.array([sin(curr_yaw), -cos(curr_yaw)])
+            cross_track_error = np.sign(np.dot(np.array([dx, dy]), left_vec)) * closest_dist
+
         yaw_error = normalise_angle(path_yaw - curr_yaw)
 
-        # 5) Standard Stanley terms
-        cross_term = atan2(self.k * cross_track_error, self.k_soft + speed)
-        desired_steering_angle = yaw_error + cross_term
-
-        desired_speed = self.desired_speed
+        desired_speed = abs(self.desired_speed)
         feedforward_accel = 0.0
 
-        if self.trajectory and self.desired_speed_source in ['path', 'trajectory']:
-            if len(self.trajectory.points) < 2 or self.current_path_parameter >= self.path.domain()[1]:
-                # End of trajectory -> stop
-                if component:
-                    component.debug_event("Stanley: Past the end of trajectory, stopping.")
-                desired_speed = 0.0
-                feedforward_accel = -2.0
-            else:
-                if self.desired_speed_source == 'path':
-                    current_trajectory_time = self.trajectory.parameter_to_time(self.current_path_parameter)
+        if self.reverse:
+            heading_term = -yaw_error
+            cross_term_input = self.k * (-cross_track_error) / (self.k_soft + abs(speed))
+            cross_term = atan2(cross_term_input, 1.0)
+            desired_steering_angle = heading_term + cross_term
+
+            if self.trajectory and self.desired_speed_source in ['path', 'trajectory']:
+                current_trajectory_param_for_eval = self.current_path_parameter
+                if hasattr(self.trajectory, 'parameter_to_time'):
+                    current_trajectory_param_for_eval = self.trajectory.parameter_to_time(self.current_path_parameter)
+
+                if self.trajectory is not None:
+                    deriv = self.trajectory.eval_derivative(current_trajectory_param_for_eval)
+                    path_speed_magnitude = np.linalg.norm(deriv)
                 else:
-                    self.current_traj_parameter += dt
-                    current_trajectory_time = self.current_traj_parameter
+                    path_speed_magnitude = 0.0
 
-                deriv = self.trajectory.eval_derivative(current_trajectory_time)
-                desired_speed = min(np.linalg.norm(deriv), self.speed_limit)
+                desired_speed = min(path_speed_magnitude, self.reverse_speed_limit)
 
+            desired_speed = desired_speed * -1.0
+
+            is_at_start_backward = self.current_path_parameter <= self.path.domain()[0] + 1e-3
+
+            if is_at_start_backward:
+                desired_speed = 0.0
+                feedforward_accel = -2.0 * -1.0
+
+            else:
                 difference_dt = 0.1
-                future_t = current_trajectory_time + difference_dt
-                if future_t > self.trajectory.domain()[1]:
-                    future_t = self.trajectory.domain()[1]
-                future_deriv = self.trajectory.eval_derivative(future_t)
-                next_desired_speed = min(np.linalg.norm(future_deriv), self.speed_limit)
+                current_speed_abs = abs(speed)
+                if current_speed_abs < 0.1: estimated_path_step = 0.1 * difference_dt * -1.0
+                else: estimated_path_step = current_speed_abs * difference_dt * -1.0
+
+                future_parameter = self.current_path_parameter + estimated_path_step
+
+                future_parameter = np.clip(future_parameter, self.path.domain()[0], self.path.domain()[-1])
+
+                future_trajectory_param_for_eval = future_parameter
+                if hasattr(self.trajectory, 'parameter_to_time'):
+                    future_trajectory_param_for_eval = self.trajectory.parameter_to_time(future_parameter)
+
+                if self.trajectory is not None:
+                    future_deriv = self.trajectory.eval_derivative(future_trajectory_param_for_eval)
+                    next_path_speed_magnitude = min(np.linalg.norm(future_deriv), self.speed_limit)
+                else:
+                    next_path_speed_magnitude = 0.0
+
+                next_desired_speed = next_path_speed_magnitude * -1.0
+
                 feedforward_accel = (next_desired_speed - desired_speed) / difference_dt
                 feedforward_accel = np.clip(feedforward_accel, -self.max_decel, self.max_accel)
+
+            desired_speed_magnitude_slowed = abs(desired_speed) * np.exp(-abs(cross_track_error) * 0.6)
+            desired_speed = desired_speed_magnitude_slowed * -1.0 if desired_speed_magnitude_slowed != 0 else 0.0
+
         else:
-            if desired_speed is None:
-                desired_speed = 4.0
+            cross_term = atan2(self.k * cross_track_error, self.k_soft + speed)
+            desired_steering_angle = yaw_error + cross_term
 
-            # Cross-track-based slowdown (less aggressive than before).
-            desired_speed *= np.exp(-abs(cross_track_error) * 0.6)
+            if self.trajectory and self.desired_speed_source in ['path', 'trajectory']:
+                if len(self.trajectory.points) < 2 or self.current_path_parameter >= self.path.domain()[1]:
+                    if component:
+                        component.debug_event("Stanley: Past the end of trajectory, stopping.")
+                    desired_speed = 0.0
+                    feedforward_accel = -2.0
+                else:
+                    if self.desired_speed_source == 'path':
+                        current_trajectory_time = self.trajectory.parameter_to_time(self.current_path_parameter)
+                    else:
+                        self.current_traj_parameter += dt
+                        current_trajectory_time = self.current_traj_parameter
 
-        # Clip to speed limit
-        if desired_speed > self.speed_limit:
+                    deriv = self.trajectory.eval_derivative(current_trajectory_time)
+                    desired_speed = min(np.linalg.norm(deriv), self.speed_limit)
+
+                    difference_dt = 0.1
+                    future_t = current_trajectory_time + difference_dt
+                    if future_t > self.trajectory.domain()[1]:
+                        future_t = self.trajectory.domain()[1]
+                    future_deriv = self.trajectory.eval_derivative(future_t)
+                    next_desired_speed = min(np.linalg.norm(future_deriv), self.speed_limit)
+                    feedforward_accel = (next_desired_speed - desired_speed) / difference_dt
+                    feedforward_accel = np.clip(feedforward_accel, -self.max_decel, self.max_accel)
+            else:
+                if desired_speed is None:
+                    desired_speed = 4.0
+
+                desired_speed *= np.exp(-abs(cross_track_error) * 0.6)
+
+        if self.reverse and abs(desired_speed) > self.reverse_speed_limit:
+            desired_speed = self.reverse_speed_limit * -1.0 if desired_speed != 0 else 0.0
+        elif not self.reverse and desired_speed > self.speed_limit:
             desired_speed = self.speed_limit
 
-        # PID for longitudinal control
         speed_error = desired_speed - speed
         output_accel = self.pid_speed.advance(e=speed_error, t=t, feedforward_term=feedforward_accel)
-
-        # Clip acceleration
         if output_accel > self.max_accel:
             output_accel = self.max_accel
         elif output_accel < -self.max_decel:
             output_accel = -self.max_decel
 
-        # Avoid negative accel when fully stopped
         if desired_speed == 0 and abs(speed) < 0.01 and output_accel < 0:
             output_accel = 0.0
 
-        # Debug
         if component is not None:
-            # component.debug("Stanley: fx, fy", (fx, fy))
             component.debug('curr pt',(curr_x,curr_y))
-            component.debug("desired_x",desired_x)
-            component.debug("desired_y",desired_y)
+            component.debug("desired_x",target_x)
+            component.debug("desired_y",target_y)
             component.debug("Stanley: path param", self.current_path_parameter)
             component.debug("Stanley: crosstrack dist", closest_dist)
             component.debug("crosstrack error", cross_track_error)
@@ -243,12 +395,6 @@ class Stanley(object):
             component.debug("Stanley: feedforward_accel (m/s^2)", feedforward_accel)
             component.debug("Stanley: output_accel (m/s^2)", output_accel)
 
-        if output_accel > self.max_accel:
-            output_accel = self.max_accel
-
-        if output_accel < -self.max_decel:
-            output_accel = -self.max_decel
-            
         self.t_last = t
         return (output_accel, desired_steering_angle)
 
@@ -256,48 +402,25 @@ class Stanley(object):
 # 3. Tracker component
 #####################################
 class StanleyTrajectoryTracker(Component):
-    """
-    A trajectory-tracking Component that uses the above Stanley controller
-    for lateral control plus PID-based longitudinal control.
-    It now includes measures to mitigate oscillations.
-    """
 
     def __init__(self, vehicle_interface=None, **kwargs):
-        """
-        :param vehicle_interface: The low-level interface to send commands to the vehicle.
-        :param kwargs: Optional parameters to pass into the Stanley(...) constructor.
-        """
         self.stanley = Stanley(**kwargs)
         self.vehicle_interface = vehicle_interface
-
+        
     def rate(self):
-        """Control frequency in Hz."""
         return 50.0
 
     def state_inputs(self):
-        """
-        Required state inputs:
-        - Vehicle state
-        - Trajectory
-        """
         return ["vehicle", "trajectory"]
 
     def state_outputs(self):
-        """No direct output state here."""
         return []
 
     def update(self, vehicle: VehicleState, trajectory: Trajectory):
-        """
-        Per control cycle:
-          1) Set the path/trajectory
-          2) Compute (acceleration, front wheel angle)
-          3) Convert front wheel angle to steering wheel angle (if necessary)
-          4) Send command to the vehicle
-        """
         self.stanley.set_path(trajectory)
+
         accel, f_delta = self.stanley.compute(vehicle, self)
 
-        # If your low-level interface expects steering wheel angle:
         steering_angle = front2steer(f_delta)
         steering_angle = np.clip(
             steering_angle,
